refactor: extend in new component
