+++ src/components/ui/extended/separator.tsx
@@ -1,0 +1,25 @@
+import { PropsWithChildren } from 'react'
+
+import { Separator as BaseSeparator } from '../separator'
+
+import { cn } from '../../../lib/utils'
+
+export function Separator({
+  children,
+  className,
+}: PropsWithChildren<{
+  className?: string
+}>) {
+  return (
+    <BaseSeparator className="flex justify-center relative">
+      <div
+        className={cn(
+          'absolute bg-background font-semibold px-1.5 text-muted-foreground text-sm -top-2.5',
+          className
+        )}
+      >
+        {children}
+      </div>
+    </BaseSeparator>
+  )
+}
